ci: drop support for python3.8
The proper version of weasyprint seems unavailable for 3.8. Nuke the
official support for it, in favor of introducing support of 3.12.

To run this app on 3.8, run the tests with ``make test`` first to see if
it fails and fix appropriately if it does so. Currently it fails due to
proper weasyprint version not being available. Needs 63.0 or higher,
available 61.2.